Project/requirements.txt
# Web Framework
streamlit==1.29.0

# LangChain Ecosystem
langchain==0.1.0
langchain-core==0.1.10
langchain-community==0.0.10
langgraph==0.0.20

# Vector Database
chromadb==0.4.22

# Embedding and ML Models
sentence-transformers==2.2.2
torch>=2.0.0
transformers>=4.35.0

# Google AI
google-generativeai==0.3.2

# Environment Management
python-dotenv==1.0.0

# Progress Bars
tqdm==4.66.1

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# HTTP Requests (for future web scraping)
requests>=2.31.0
beautifulsoup4>=4.12.0

# Utilities
pydantic>=2.5.0
typing-extensions>=4.8.0

# Development Tools (optional)
pytest>=7.4.0
black>=23.11.0
flake8>=6.1.0

# Logging
colorlog>=6.8.0

